Trying to solve a problem with the compiler x86_64-w64-mingw32-g++

+++ src/brace_cleanup.cpp
@@ -689,8 +689,8 @@ static void parse_cleanup(parse_frame_t *frm, chunk_t *pc)
          if (!cpd.unc_off_used)
          {
             /* fatal error */
-            fprintf(stderr, "Unmatched BRACE_CLOSE\nat line=%zu, column=%zu\n",
-                    pc->orig_line, pc->orig_col);
+            fprintf(stderr, "Unmatched BRACE_CLOSE\nat line=%d, column=%d\n",
+                    (int)pc->orig_line, (int)pc->orig_col);
             exit(EXIT_FAILURE);
          }
       }
